refactor(server): migrate shared helpers to canonical core paths
Move server cross-cutting helpers and their tests from `packages/server/shared/*`
to `packages/server/core/*`, and update all server imports/exports accordingly.
Also update server Vitest coverage include paths and align architecture docs to
the canonical `core/*` helper structure.

No runtime behavior changes intended.

Files:
- ARCHITECTURE.md
- packages/server/index.ts
- packages/server/vitest.config.ts
- packages/server/core/date-only.ts
- packages/server/core/logger.ts
- packages/server/core/result.ts
- packages/server/core/types.ts
- packages/server/core/__tests__/date-only.test.ts
- packages/server/core/__tests__/logger.test.ts
- packages/server/core/__tests__/types.test.ts
- packages/server/admin/core/service.ts
- packages/server/admin/dogs/lookups/list-breeder-options.ts
- packages/server/admin/dogs/lookups/list-owner-options.ts
- packages/server/admin/dogs/lookups/list-parent-options.ts
- packages/server/admin/dogs/manage/create-dog.ts
- packages/server/admin/dogs/manage/delete-dog.ts
- packages/server/admin/dogs/manage/list-dogs.ts
- packages/server/admin/dogs/manage/update-dog.ts
- packages/server/admin/users/manage/create-user.ts
- packages/server/admin/users/manage/delete-user.ts
- packages/server/admin/users/manage/list-users.ts
- packages/server/admin/users/manage/set-user-password.ts
- packages/server/admin/users/manage/set-user-status.ts
- packages/server/dogs/newest/get-newest-beagle-dogs.ts
- packages/server/dogs/profile/get-beagle-dog-profile.ts
- packages/server/dogs/search/service.ts
- packages/server/home/statistics/get-home-statistics.ts
- packages/server/imports/date-key.ts
- packages/server/imports/service.ts

Ref BEJ-17
