+++ packages/server/admin/dogs/lookups/list-breeder-options.ts
@@ -4,8 +4,8 @@ import type {
   AdminDogLookupRequest,
   CurrentUserDto,
 } from "@beagle/contracts";
-import { toErrorLog, withLogContext } from "../../../shared/logger";
-import type { ServiceResult } from "../../../shared/result";
+import { toErrorLog, withLogContext } from "../../../core/logger";
+import type { ServiceResult } from "../../../core/result";
 import { requireAdmin } from "../../core/service";
 import { normalizeQuery, parseLookupLimit } from "../manage/normalization";
