+++ packages/server/admin/dogs/manage/create-dog.ts
@@ -8,8 +8,8 @@ import type {
   CreateAdminDogRequest,
   CreateAdminDogResponse,
 } from "@beagle/contracts";
-import { toErrorLog, withLogContext } from "../../../shared/logger";
-import type { ServiceResult } from "../../../shared/result";
+import { toErrorLog, withLogContext } from "../../../core/logger";
+import type { ServiceResult } from "../../../core/result";
 import {
   hasMaxLength,
   isValidRegistrationNo,
