+++ packages/server/admin/dogs/manage/delete-dog.ts
@@ -7,8 +7,8 @@ import type {
   DeleteAdminDogRequest,
   DeleteAdminDogResponse,
 } from "@beagle/contracts";
-import { toErrorLog, withLogContext } from "../../../shared/logger";
-import type { ServiceResult } from "../../../shared/result";
+import { toErrorLog, withLogContext } from "../../../core/logger";
+import type { ServiceResult } from "../../../core/result";
 import { normalizeRequiredText } from "./normalization";
 
 export async function deleteAdminDog(
