+++ packages/server/imports/date-key.ts
@@ -1,4 +1,4 @@
-import { toBusinessDateOnly } from "../shared/date-only";
+import { toBusinessDateOnly } from "../core/date-only";
 
 export function toOwnershipDateKey(value: Date | null): string {
   // Internal dedupe key: MariaDB/MySQL unique indexes allow multiple NULL values.
