+++ packages/server/index.ts
@@ -17,5 +17,5 @@ export { betterAuth } from "./auth";
 export { createDogsService, dogsService } from "./dogs";
 export { createImportsService, importsService } from "./imports";
 export { createStatsService, statsService } from "./home";
-export type { ServiceResult } from "./shared/result";
-export { logger, toErrorLog, withLogContext } from "./shared/logger";
+export type { ServiceResult } from "./core/result";
+export { logger, toErrorLog, withLogContext } from "./core/logger";
